requirements.txt
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.1.0
autopep8==2.3.1
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.7.14
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
fqdn==1.5.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_contrib_core==0.4.2
jupyter_core==5.7.2
jupyter_latex_envs==1.4.6
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
missingno==0.5.2
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.0.1
opencv-python==4.12.0.88
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.6.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pycodestyle==2.12.1
pycparser==2.22
Pygments==2.18.0
pyodbc==5.2.0
pyparsing==3.1.2
pyreadstat==1.2.7
Pyrogram==2.0.106
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
scikit-learn==1.7.0
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.2
svgpathtools==1.7.1
svgwrite==1.4.3
sympy==1.14.0
tabula-py==2.10.0
terminado==0.18.1
TgCrypto==1.2.5
threadpoolctl==3.6.0
tinycss2==1.2.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
XlsxWriter==3.2.0
